Added code docs

+++ src/texc_expand/expand.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <stdbool.h>
+#include <stdint.h>
 
 #include "../texc_tags/tags.h"
 
@@ -8,7 +9,14 @@ typedef struct {
     // Is the expand undoable to get original text
     bool is_undoable;
     // The amount of characters typed to perform the undo
-    int typed_count;
+    size_t typed_count;
 } ExpandSettings;
 
+/**
+ * @brief Deletes the match and types the expand tag 
+ *
+ * @param delete_len The length of the match text
+ * @param tag The expand tag
+ * @return settigns related to the expand
+ */
 ExpandSettings expand_text(size_t delete_len, Tag *tag);
